pyproject.toml
[project]
name = "lightning-vocoders"
version = "0.1.0"
description = "Add a short description here"
authors = [
    { name = "Wataru Nakata", email = "wataru9871@gmail.com" }
]
dependencies = ["torch~=2.0.1", "soundfile~=0.12.1", "transformers~=4.29.2", "umap-learn[plot]~=0.5.3", "torchaudio~=2.0.2", "lightning~=2.0.2", "hydra-core~=1.3.2", "pyrootutils~=1.0.4", "webdataset~=0.2.48", "tensorboard~=2.13.0", "wandb~=0.15.3"]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ["ipykernel~=6.23.1", "black~=23.3.0"]
[tool.hatch.metadata]
allow-direct-references = true
